+++ test/integration/elastic_db_repo/list_dumps.py
@@ -64,11 +64,9 @@ class UnitTest(unittest.TestCase):
         self.japd = self.cfg.japd if hasattr(self.cfg, "japd") else None
         self.ca_cert = self.cfg.ssl_client_ca if hasattr(
             self.cfg, "ssl_client_ca") else None
-        self.scheme = self.cfg.scheme if hasattr(
-            self.cfg, "scheme") else "https"
         self.els = elcs.ElasticSearchRepo(
-            self.cfg.host, port=self.cfg.port, user=self.user, japd=self.japd,
-            ca_cert=self.ca_cert, scheme=self.scheme)
+            self.cfg.host, user=self.user, japd=self.japd,
+            ca_cert=self.ca_cert)
         self.els.connect()
 
         if self.els.repo_dict:
@@ -102,9 +100,8 @@ class UnitTest(unittest.TestCase):
         """
 
         els = elcs.ElasticSearchDump(
-            self.cfg.host, port=self.cfg.port, repo=self.repo_name,
-            user=self.user, japd=self.japd, ca_cert=self.ca_cert,
-            scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name, user=self.user, japd=self.japd,
+            ca_cert=self.ca_cert)
         els.connect()
         err_flag, msg = els.dump_db()
 
@@ -128,9 +125,8 @@ class UnitTest(unittest.TestCase):
         """
 
         els = elcs.ElasticSearchDump(
-            self.cfg.host, port=self.cfg.port, repo=self.repo_name,
-            user=self.user, japd=self.japd, ca_cert=self.ca_cert,
-            scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name, user=self.user, japd=self.japd,
+            ca_cert=self.ca_cert)
         els.connect()
         err_flag, msg = els.dump_db()